fix(generator): handle ptrdiff_t type correctly in C type mapping
Adds special case handling for ptrdiff_t in getCType function to preserve its
type name rather than converting it to a Go equivalent.

This fixes type mismatch errors on macOS where ptrdiff_t is a distinct type
from int64_t despite having the same size.

+++ functions.gen.go
@@ -14172,7 +14172,7 @@ func AVImageCopyPlane(dst unsafe.Pointer, dstLinesize int, src unsafe.Pointer, s
         size (i.e. 64) to get improved performance.
 */
 func AVImageCopyPlaneUcFrom(dst unsafe.Pointer, dstLinesize int64, src unsafe.Pointer, srcLinesize int64, bytewidth int64, height int) {
-	C.av_image_copy_plane_uc_from((*C.uint8_t)(dst), C.int64_t(dstLinesize), (*C.uint8_t)(src), C.int64_t(srcLinesize), C.int64_t(bytewidth), C.int(height))
+	C.av_image_copy_plane_uc_from((*C.uint8_t)(dst), C.ptrdiff_t(dstLinesize), (*C.uint8_t)(src), C.ptrdiff_t(srcLinesize), C.ptrdiff_t(bytewidth), C.int(height))
 }
 
 // --- Function av_image_copy ---
